requirements.txt

absl-py==1.3.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
black==22.3.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
cython==0.29.32
docker-pycreds==0.4.0
fairscale==0.4.13
filelock==3.9.0
fonttools==4.38.0
future==0.18.2
fvcore==0.1.5.post20221221
gitdb==4.0.10
gitpython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.11.1
hydra-core==1.3.1
idna==3.4
imageio==2.23.0
importlib-metadata==6.0.0
iopath==0.1.9
joblib==1.2.0
kiwisolver==1.4.4
markdown==3.4.1
markupsafe==2.1.1
matplotlib==3.6.2
mypy-extensions==0.4.3
networkx==2.8.8
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.68
packaging==22.0
pandas==1.5.2
pathspec==0.10.3
pathtools==0.1.2
pillow==9.3.0
protobuf==3.20.3
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pydot==1.4.2
pyopenssl==22.0.0
pyparsing==3.0.9
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
pywavelets==1.4.1
pyyaml==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
sentry-sdk==1.12.1
setproctitle==1.3.2
shapely==2.0.0
six==1.16.0
stack_data==0.2.0
submitit==1.4.5
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.12
tomli==2.0.1
torch==1.13.1
torch-tb-profiler==0.4.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.13.8
werkzeug==2.2.2
yacs==0.1.8
yapf==0.31.0
zipp==3.11.0
wget==3.2
packaging==22.0